backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.6.1
pypdf2==3.0.1
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4
aiofiles==23.2.1
httpx==0.25.2
tenacity==8.2.3
loguru==0.7.2
rich==13.7.0
requests==2.32.5
chromadb==0.5.23
sentence-transformers==2.2.2
pycryptodome==3.23.0
openpyxl==3.1.2
xlrd==2.0.1